chore: move lightning en/decodable impls to btc.rs
